django/requirements.txt

Django>=5.0
requests
google-cloud-firestore>=2.16
python-dotenv
pillow
